move formatValueToDisplay
