+++ src/components/TokenDisplay.test.tsx
@@ -3,8 +3,7 @@ import BigNumber from 'bignumber.js'
 import * as React from 'react'
 import 'react-native'
 import { Provider } from 'react-redux'
-import TokenDisplay from 'src/components/TokenDisplay'
-import { formatValueToDisplay } from 'src/components/LegacyTokenDisplay'
+import TokenDisplay, { formatValueToDisplay } from 'src/components/TokenDisplay'
 import { LocalCurrencyCode } from 'src/localCurrency/consts'
 import { RootState } from 'src/redux/reducers'
 import { createMockStore, getElementText, RecursivePartial } from 'test/utils'
